Move nano::pending_info and nano::pending_key in to their own file.
