+++ nano/node/bootstrap/bootstrap_bulk_pull.hpp
@@ -2,6 +2,7 @@
 
 #include <nano/node/messages.hpp>
 #include <nano/node/transport/socket.hpp>
+#include <nano/secure/pending_info.hpp>
 
 #include <unordered_set>
