+++ nano/secure/common.cpp
@@ -218,72 +218,6 @@ nano::keypair::keypair (std::string const & prv_a)
 	ed25519_publickey (prv.bytes.data (), pub.bytes.data ());
 }
 
-nano::pending_info::pending_info (nano::account const & source_a, nano::amount const & amount_a, nano::epoch epoch_a) :
-	source (source_a),
-	amount (amount_a),
-	epoch (epoch_a)
-{
-}
-
-bool nano::pending_info::deserialize (nano::stream & stream_a)
-{
-	auto error (false);
-	try
-	{
-		nano::read (stream_a, source.bytes);
-		nano::read (stream_a, amount.bytes);
-		nano::read (stream_a, epoch);
-	}
-	catch (std::runtime_error const &)
-	{
-		error = true;
-	}
-
-	return error;
-}
-
-size_t nano::pending_info::db_size () const
-{
-	return sizeof (source) + sizeof (amount) + sizeof (epoch);
-}
-
-bool nano::pending_info::operator== (nano::pending_info const & other_a) const
-{
-	return source == other_a.source && amount == other_a.amount && epoch == other_a.epoch;
-}
-
-nano::pending_key::pending_key (nano::account const & account_a, nano::block_hash const & hash_a) :
-	account (account_a),
-	hash (hash_a)
-{
-}
-
-bool nano::pending_key::deserialize (nano::stream & stream_a)
-{
-	auto error (false);
-	try
-	{
-		nano::read (stream_a, account.bytes);
-		nano::read (stream_a, hash.bytes);
-	}
-	catch (std::runtime_error const &)
-	{
-		error = true;
-	}
-
-	return error;
-}
-
-bool nano::pending_key::operator== (nano::pending_key const & other_a) const
-{
-	return account == other_a.account && hash == other_a.hash;
-}
-
-nano::account const & nano::pending_key::key () const
-{
-	return account;
-}
-
 nano::unchecked_info::unchecked_info (std::shared_ptr<nano::block> const & block_a) :
 	block (block_a),
 	modified_m (nano::seconds_since_epoch ())
