+++ nano/secure/common.hpp
@@ -98,33 +98,6 @@ public:
 	nano::raw_key prv;
 };
 
-/**
- * Information on an uncollected send
- */
-class pending_info final
-{
-public:
-	pending_info () = default;
-	pending_info (nano::account const &, nano::amount const &, nano::epoch);
-	size_t db_size () const;
-	bool deserialize (nano::stream &);
-	bool operator== (nano::pending_info const &) const;
-	nano::account source{};
-	nano::amount amount{ 0 };
-	nano::epoch epoch{ nano::epoch::epoch_0 };
-};
-class pending_key final
-{
-public:
-	pending_key () = default;
-	pending_key (nano::account const &, nano::block_hash const &);
-	bool deserialize (nano::stream &);
-	bool operator== (nano::pending_key const &) const;
-	nano::account const & key () const;
-	nano::account account{};
-	nano::block_hash hash{ 0 };
-};
-
 class endpoint_key final
 {
 public:
