+++ nano/secure/pending_info.hpp
@@ -1,0 +1,35 @@
+#pragma once
+
+#include <nano/lib/epoch.hpp>
+#include <nano/lib/numbers.hpp>
+#include <nano/lib/stream.hpp>
+
+namespace nano
+{
+/**
+ * Information on an uncollected send
+ */
+class pending_info final
+{
+public:
+	pending_info () = default;
+	pending_info (nano::account const &, nano::amount const &, nano::epoch);
+	size_t db_size () const;
+	bool deserialize (nano::stream &);
+	bool operator== (nano::pending_info const &) const;
+	nano::account source{};
+	nano::amount amount{ 0 };
+	nano::epoch epoch{ nano::epoch::epoch_0 };
+};
+class pending_key final
+{
+public:
+	pending_key () = default;
+	pending_key (nano::account const &, nano::block_hash const &);
+	bool deserialize (nano::stream &);
+	bool operator== (nano::pending_key const &) const;
+	nano::account const & key () const;
+	nano::account account{};
+	nano::block_hash hash{ 0 };
+};
+} // namespace nano
